+++ packages/bylaw_credo/.credo.exs
@@ -147,6 +147,7 @@ end
           {Bylaw.Credo.Check.Elixir.NoThen, []},
           {Bylaw.Credo.Check.Ecto.PipeBasedQueries, false},
           {Bylaw.Credo.Check.Elixir.PreferEmptyListChecks, false},
+          {Bylaw.Credo.Check.Elixir.PreferBlockIf, []},
           {Bylaw.Credo.Check.Elixir.PreferEnumCount, []},
           {Bylaw.Credo.Check.Elixir.PreferEnumUniqBy, []},
           {Bylaw.Credo.Check.Elixir.PreferListTypeSyntax, []},